add BlockFactoryDenebTest
